refactor(render)
- move AddTorusQuad, getAngle & getNextAngle into VertexBuilder
- replace coords with entity snapshot (coords, yaw & pitch rot, bbHeight)
- create bool for toggle halo in Soul
- move fadeProgress (fade) into Mode
