+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/hats/modes/HatsCone.kt
@@ -26,6 +26,8 @@ import net.ccbluex.liquidbounce.render.ClientRenderPipelines
 import net.ccbluex.liquidbounce.render.WorldRenderEnvironment
 import net.ccbluex.liquidbounce.render.drawCustomMesh
 import net.ccbluex.liquidbounce.render.engine.type.Color4b
+import net.ccbluex.liquidbounce.render.getAngle
+import net.ccbluex.liquidbounce.render.getNextAngle
 import net.ccbluex.liquidbounce.render.setColor
 import kotlin.math.cos
 import kotlin.math.sin
