+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/hats/modes/HatsFlower.kt
@@ -24,8 +24,11 @@ import net.ccbluex.liquidbounce.features.module.modules.render.hats.HatsColorSet
 import net.ccbluex.liquidbounce.features.module.modules.render.hats.HatsMode
 import net.ccbluex.liquidbounce.render.ClientRenderPipelines
 import net.ccbluex.liquidbounce.render.WorldRenderEnvironment
+import net.ccbluex.liquidbounce.render.addTorusQuad
 import net.ccbluex.liquidbounce.render.drawCustomMesh
 import net.ccbluex.liquidbounce.render.engine.type.Color4b
+import net.ccbluex.liquidbounce.render.getAngle
+import net.ccbluex.liquidbounce.render.getNextAngle
 import net.minecraft.util.Mth
 import kotlin.math.abs
