+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/totemeffect/TotemEffectMode.kt
@@ -23,15 +23,15 @@ import net.ccbluex.liquidbounce.config.types.group.Mode
 import net.ccbluex.liquidbounce.config.types.group.ModeValueGroup
 import net.ccbluex.liquidbounce.event.events.WorldRenderEvent
 import net.ccbluex.liquidbounce.event.handler
-import net.ccbluex.liquidbounce.features.module.modules.render.totemeffect.ModuleTotemEffect.coords
+import net.ccbluex.liquidbounce.features.module.modules.render.totemeffect.ModuleTotemEffect.TotemPopSnapshot
+import net.ccbluex.liquidbounce.features.module.modules.render.totemeffect.ModuleTotemEffect.entities
 import net.ccbluex.liquidbounce.render.WorldRenderEnvironment
 import net.ccbluex.liquidbounce.render.renderEnvironment
 import net.ccbluex.liquidbounce.render.withPositionRelativeToCamera
 import net.ccbluex.liquidbounce.render.withPush
 import net.ccbluex.liquidbounce.utils.math.Easing
-import net.minecraft.world.phys.Vec3
 
-abstract class TotemEffectMode(name:String) : Mode(name) {
+abstract class TotemEffectMode(name: String) : Mode(name) {
     final override val parent: ModeValueGroup<*>
         get() = ModuleTotemEffect.modes
 
@@ -40,26 +40,24 @@ abstract class TotemEffectMode(name:String) : Mode(name) {
     protected val animCurve by easing("AnimCurve", Easing.EXPONENTIAL_OUT)
     protected val canBeCovered by boolean("CanBeCovered", false)
 
-    protected abstract fun WorldRenderEnvironment.drawTotemEffect(
-        progress: Float,
-        age: Float,
-        pos: Vec3,
-        event: WorldRenderEvent
-    )
+    protected abstract fun WorldRenderEnvironment
+        .drawTotemEffect(progress: Float, age: Float, entity: TotemPopSnapshot, fade: Float, event: WorldRenderEvent)
 
     @Suppress("unused")
     private val renderHandler = handler<WorldRenderEvent> { event ->
         event.renderEnvironment {
-            coords.forEach {
-                withPositionRelativeToCamera(it.value) {
-                    val age = lifetime - coords.timeToDie(it) + event.partialTicks
+            entities.forEach {
+                withPositionRelativeToCamera(it.value.pos.add(0.0, it.value.bbHeight / 2.0, 0.0)) {
+                    val age = lifetime - entities.timeToDie(it) + event.partialTicks
 
                     val progress = animCurve
                         .transform(age / lifetime)
                         .coerceIn(0f, 1f)
 
+                    val fade = if (progress >= fade) (progress - fade) / (1f - fade).coerceAtLeast(0.001f) else 0f
+
                     poseStack.withPush {
-                        drawTotemEffect(progress, age, it.value, event)
+                        drawTotemEffect(progress, age, it.value, fade, event)
                     }
                 }
             }
