+++ src/main/kotlin/net/ccbluex/liquidbounce/render/VertexBuilder.kt
@@ -26,12 +26,15 @@ import com.mojang.blaze3d.vertex.PoseStack
 import com.mojang.blaze3d.vertex.VertexConsumer
 import net.ccbluex.liquidbounce.render.engine.type.Color4b
 import net.ccbluex.liquidbounce.render.engine.type.Vec3f
+import net.ccbluex.liquidbounce.utils.math.fastCos
+import net.ccbluex.liquidbounce.utils.math.fastSin
 import net.ccbluex.liquidbounce.utils.math.forAllFaces
 import net.ccbluex.liquidbounce.utils.math.forAllSideFaces
 import net.ccbluex.liquidbounce.utils.math.forAllSideOutlineEdges
 import net.ccbluex.liquidbounce.utils.render.begin
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
+import net.minecraft.util.Mth
 import net.minecraft.world.phys.AABB
 import net.minecraft.world.phys.Vec3
 import net.minecraft.world.phys.shapes.VoxelShape
@@ -222,6 +225,60 @@ private fun VertexConsumer.addColoredVertex(
     if (color != null) {
         setColor(color.argb)
     }
+}
+
+fun getAngle(i: Int, segments: Int) = i * Mth.TWO_PI / segments
+fun getNextAngle(i: Int, segments: Int) = (i + 1) * Mth.TWO_PI / segments
+
+fun VertexConsumer.addTorusQuad(
+    pose: PoseStack.Pose,
+    innerSegments: Int,
+    outerCurAngle: Float,
+    outerNextAngle: Float,
+    outerCurRadius: Float,
+    outerNextRadius: Float,
+    innerRadius: Float,
+    innerI: Int,
+    color: Color4b,
+) {
+    val innerCurAngle = getAngle(innerI, innerSegments)
+    val innerNextAngle = getNextAngle(innerI, innerSegments)
+
+    val curMainSin = outerCurAngle.fastSin()
+    val curMainCos = outerCurAngle.fastCos()
+    val nextMainSin = outerNextAngle.fastSin()
+    val nextMainCos = outerNextAngle.fastCos()
+
+    val innerCurSin = innerCurAngle.fastSin()
+    val innerCurCos = innerCurAngle.fastCos()
+    val innerNextSin = innerNextAngle.fastSin()
+    val innerNextCos = innerNextAngle.fastCos()
+
+    val curTubeY = innerRadius * innerCurSin
+    val nextTubeY = innerRadius * innerNextSin
+    val curTubeOffset = innerRadius * innerCurCos
+    val nextTubeOffset = innerRadius * innerNextCos
+
+    val p1Radius = outerCurRadius + curTubeOffset
+    val p2Radius = outerCurRadius + nextTubeOffset
+    val p3Radius = outerNextRadius + curTubeOffset
+    val p4Radius = outerNextRadius + nextTubeOffset
+
+    val p1x = p1Radius * curMainSin
+    val p1z = p1Radius * curMainCos
+    val p2x = p2Radius * curMainSin
+    val p2z = p2Radius * curMainCos
+    val p3x = p3Radius * nextMainSin
+    val p3z = p3Radius * nextMainCos
+    val p4x = p4Radius * nextMainSin
+    val p4z = p4Radius * nextMainCos
+
+    addVertex(pose, p1x, curTubeY, p1z).setColor(color)
+    addVertex(pose, p2x, nextTubeY, p2z).setColor(color)
+    addVertex(pose, p3x, curTubeY, p3z).setColor(color)
+    addVertex(pose, p2x, nextTubeY, p2z).setColor(color)
+    addVertex(pose, p4x, nextTubeY, p4z).setColor(color)
+    addVertex(pose, p3x, curTubeY, p3z).setColor(color)
 }
 
 /**
